+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `useLoadDataState()` and `useLoadResourcesState()` to always pass an `AbortSignal`
+
 ## [5.1.1] - 2023-09-07
 
 ### Changed